+++ ChangeLog.md
@@ -3,6 +3,9 @@ XP Reflection ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Added `lang.reflect.Methods::with()` method to enumerate all methods
+  with a given annotation
+  (@thekid)
 * Fixed accessing meta data for type members via `xp::$meta` cache
   (@thekid)